IMPORTANT FIX
Fix bug with convert teacher and subject entity to domain

+++ src/main/kotlin/com/prinzh/schedule/app/common/extension/StringExtensions.kt
@@ -25,7 +25,7 @@ fun String.toColor(): String {
     val color = this.replaceFirst("#", "")
 
     try {
-        color.toInt(16)
+        color.toLong(16)
     } catch (t: Throwable) {
         throw BadRequestException("Invalid credentials")
     }
